backend/requirements.txt
fastapi[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
python-dotenv
httpx
uvicorn[standard]
python-multipart
aiofiles
numpy
Pillow
transformers
webrtcvad-wheels
soundfile
pytest
pytest-asyncio
